+++ src/components/task/TaskListItem.jsx
@@ -162,6 +162,35 @@ export default function TaskListItem({ task, rowNumber = null }) {
 		}
 	}
 
+	/**
+	 * Filters whether the toggle column should be rendered
+	 * in the TaskListItem.
+	 *
+	 * @since [unreleased]
+	 *
+	 * @param {boolean} renderToggle If to render the toggle column.
+	 * @param {Object}  task The represented task.
+	 */
+	renderToggle = window.Completionist.hooks.applyFilters(
+		'TaskListItem_if_render_toggle',
+		renderToggle,
+		task
+	);
+
+	/**
+	 * Filters whether the TaskListItem can be toggled (expanded).
+	 *
+	 * @since [unreleased]
+	 *
+	 * @param {boolean} allowToggle If to allow toggling (expansion).
+	 * @param {Object}  task The represented task.
+	 */
+	allowToggle = window.Completionist.hooks.applyFilters(
+		'TaskListItem_if_allow_toggle',
+		allowToggle,
+		task
+	);
+
 	let maybeToggle = null;
 	if ( renderToggle ) {
 		let maybeToggleIcon = null;
@@ -174,7 +203,28 @@ export default function TaskListItem({ task, rowNumber = null }) {
 
 	let maybeExpandedDetails = null;
 	if ( isExpanded ) {
+
 		extraClassNames += ' --is-expanded';
+
+		/**
+		 * Filters content to display after the task
+		 * description within a TaskListItem.
+		 *
+		 * Note that a task does not require a description
+		 * for this content to be displayed.
+		 *
+		 * @since [unreleased]
+		 *
+		 * @param {Object[]} content An array of React
+		 * components or JSX elements to be rendered.
+		 * @param {Object}   task The represented task.
+		 */
+		const maybeContentAfterDescription = window.Completionist.hooks.applyFilters(
+			'TaskListItem_content_after_description',
+			[],
+			task
+		);
+
 		maybeExpandedDetails = (
 			<div className="expanded-details">
 				{ rowNumber && <div className="spacer row-number"></div> }
@@ -183,6 +233,7 @@ export default function TaskListItem({ task, rowNumber = null }) {
 				<div className="details">
 					{maybeTags}
 					{maybeDescription}
+					{maybeContentAfterDescription}
 					{maybeSubtaskList}
 					{maybeAttachments}
 				</div>